quant-daily-report/requirements.txt
# requirements.txt - 生产环境依赖

# 核心数据处理
pandas>=2.2.0
numpy>=1.26.3
python-dateutil>=2.8.2
pytz>=2023.3.post1

# A股数据采集
akshare>=1.10.45
tushare>=1.2.91
baostock>=0.8.8

# HTTP请求与解析
requests>=2.31.0
aiohttp>=3.9.1
brotli>=1.1.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
feedparser>=6.0.10

# 技术指标计算
TA-Lib>=0.4.29
pandas-ta>=0.3.14b0
technical>=0.1.4
scipy>=1.11.4

# 量化回测
pyqlib

# 可视化与报告
mplfinance>=0.12.10b0
plotly>=5.18.0
matplotlib>=3.8.2
seaborn>=0.13.0
jinja2>=3.1.3
weasyprint>=59.0
markdown>=3.5.1

# 自动化部署
pygithub>=2.3.0
python-dotenv>=1.0.0
schedule>=1.2.1
apscheduler>=3.10.4
pyyaml>=6.0.1
toml>=0.10.2

# 日志与监控
loguru>=0.7.2
sentry-sdk>=1.40.0

# 其他工具
click>=8.1.7
colorama>=0.4.6
tqdm>=4.66.1
joblib>=1.3.2